pyproject.toml
[project]
name = "dbt-contracts"
version = "0.1.0"
description = "Contract-driven dbt workflow using ODPS and ODCS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "logfire>=4.16.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "questionary>=2.0",
    "open-data-contract-standard>=2.0.0",
    "datacontract-cli>=0.10",
]

[project.scripts]
dbt-contracts = "dbt_contracts.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.12",
    "pre-commit>=4.5.1",
    "zensical>=0.0.19",
    "pytest>=9.0.2",
    "pytest-mock>=3.0",
    "mkdocstrings-python>=2.0.1",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

# Enforce docstrings to generate documenation from

[tool.ruff.lint]
select = ["D", "E4", "E7", "E9", "F"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
